refactor: 가격 관련 데이터 타입 Double에서 Int로 변경

+++ O-G-Market/Models/Response/PostingInfo.swift
@@ -13,9 +13,9 @@ struct PostingInfo: Decodable {
     let name: String
     let thumbnail: String
     let currency: Currency
-    let price: Double
-    let bargainPrice: Double
-    let discountedPrice: Double
+    let price: Int
+    let bargainPrice: Int
+    let discountedPrice: Int
     let stock: Int
     let createdAt: String
     let issuedAt: String
